+++ source/Utilities/LibMultiSense/VersionInfoUtility/VersionInfoUtility.cc
@@ -79,7 +79,7 @@ int main(int argc, char** argv)
         switch(c)
         {
             case 'a': ip_address = std::string(optarg); break;
-            case 'm': mtu = atoi(optarg); break;
+            case 'm': mtu = static_cast<uint16_t>(atoi(optarg)); break;
             default: usage(*argv); break;
         }
     }
